Add APIs for pipeline assessment
Added a system service that will be used to store pipeline drafts
and perform assessments on those drafts.

Refactored some classes into a delta-proto module that contains
objects used in the RESTful APIs. Also introduced assessments to
the API module, and modified the source and target APIs. Sources
are now responsible for generating a list of tables and fetching
details about each table. Targets are now responsible for assessing
whether there will be problems replicating a given table.