implement assessment endpoints
Implement service endpoints that create a pipeline assessment and
table assessment for a draft.

Introduced APIs to the DeltaSource for converting a raw list of
SQL column types into a standard CDAP Schema that will be sent
to the DeltaTarget. Also added an API for the source to assess a
table so that we can capture partial column support, such as a
db specific type that doesn't have full support in CDAP. For example,
some datetime types will lose granularity, so they are partially
supported.

Added a tables list to the DeltaConfig to configure which tables
the source should read from.