cleanup
Renamed Support to ColumnSupport and add more javadocs

Changed TableDetail back to what it was before instead of making
the columns generic. This is so TableDetail and StandardizedTableDetail
do not share a common parent, to make it more clear what each of
them represent. Added more javadocs around what each one is.