cleanup restful APIs
Removed database and table names from the path of the describe
table and assess table endpoints. Instead, the database and table
names are passed in through the request body.

This change was made because these endpoints are actions to perform
using a draft, and not elements within a draft.
It also makes the API more extensible in case there needs to be
additional information passed in the future.