fix the table assessment endpoint response
Fixed the table assessment endpoint to return an object that
contains both the column name and type as seen by both the source
and target. Added a FullTableAssessment class to represent this
response. Also add a list of problems to the TableAssessment
class so that a target can record that it needs a primary key,
and other types of similar problems