CDAP-16244 add metrics for events applied
Added methods to the DeltaTargetContext to increment in-memory
counts for the number of DML and DDL operations replicated.
These counts are emitted as metrics after an offset is committed.

Target plugins must correctly increment these counters in order
for the metrics to be accurate. This cannot be done by the app
because the app does not have control over when the plugin
decides to commit the offset.