CDAP-16246 add event type filtering
Enhanced the config to allow the user to specify a blacklist for
DML and DDL event types. A blacklist can be set at the pipeline
level for all tables, or for a specific table.

Refactored the source API to take a new EventReaderDefinition
class that contains information about what to read. This is more
extensible, as it allows us to add more constraints without breaking
backwards compatibility in the API.