CDAP-16245 add table state apis
Added methods to the source and target contexts to allow them to
record state information. The source can record that the pipeline
is snapshotting, replicating, or failing. The target can record
that specific tables are ok or failing. If something is failing,
a replication error containing a message and a stack trace can
also be provided.

Also added a REST endpoint to get the state for a pipeline.

Also added a small fix to correctly store the sequence number
along with the offset.