CDAP-16251 retry failures up to a limit
Refactored the worker so that events emitted by the source are
first added to a work queue instead of directly passed to the
target. This is required so that the app is in control of when
calls to apply a DMLEvent or DDLEvent are made, and can properly
handle any exceptions that are thrown if they cannot be applied.

If an exception is thrown when a DML or DDL event is being applied,
the worker will stop the event reader and consumer, reset the offset
and sequence number to the last committed offset, then start the
reader from that old offset. The worker will continue to retry
until a configurable time limit is hit.

Also added some more localized retries around storing offsets
and pipeline state.