CDAP-16313 add parallelism
Added a parallelism config to the delta config that lets the user
specify either the number of worker instances to use. At deployment
time, the source tables are assigned to different worker instances
in a round robin fashion. Alternatively, the user can also manually
provide the table assignments instead of just the number of worker
instances.

Refactored the worker such that each worker instance only reads and
applies changes to its assigned set of tables. Offsets, sequence
numbers, and state were all changed to be per worker instance.

Currently, there is no way to change the number of instances
after the app is deployed, due to the tables being assigned at
deployment time and because there isn't currently a way to merge
different instance offsets/states.