CDAP-16737 protect against bad event readers
Changed the failure retry logic so that a new event queue is
used instead of re-using the existing queue. This is to protect
against bad event readers that are not really stopped after they
return from the stop() call. This prevents those types of
readers from adding events to the queue that shouldn't be there,
breaking the event ordering guarantee.

Also added some logging when readers and consumers are starting
and stopping so that the lifecycle of everything is more clear.

Also made the EventEmitter throw InterruptedException, since the
caller really does need to know when it is interrupted in
order to handle it properly and ensure everything shuts down.